feature: moved the modal window into a separate component
